Create a grid table for a reference grid

Create a grid table for a reference grid

create_grid_table(grid)

Arguments

grid

The reference grid as object of class stars.

Value

A data.table object. See details.

Details

The grid table of a GTFS-Multi feed always has the same structure as the stops table in a regular GTFS feed. Instead of a true transit stop location, each row represents a grid point (i.e. the centroid of a pixel in the reference grid).

See also