Create a transfer table for a set of origins and destinations

Create a transfer table for a set of origins and destinations

create_transfer_table(
  origins,
  destinations,
  streetnet,
  d_limit = Inf,
  min_time = 120,
  ...
)

Arguments

origins

A data.table object containing the origin locations of the transfer trips. See details.

destinations

A data.table object containing the destination locations of the transfer trips. See details.

streetnet

The street network as an object of class sfnetworks, dodgr_streetnet or dodgr_streetnet_sc.

d_limit

Upper distance limit (as-the-crow-flies) in meters to consider a possible transfer between an origin and a destination.

min_time

Minimum travel time for any transfer trip in seconds, independent of the calculated travel time by the routing algorithm.

...

Additional parameters passed on to streetnet_traveltimes(). When parameter streetnet is of class sfnetworks, these include the time_column and weight_column parameters. See streetnet_traveltimes() for details.

Details

Both origins and destinations should be given in the format of a stops table of a GTFS feed.

The format of a transfer table can be used to model both true transit transfer trips as well as access, direct and egress trips:

When modelling access trips, the origins will be grid points and the destinations transit stop locations.

When modelling egress trips the origins will be transit stop locations and the destinations grid points.

When modelling direct trips both origins and destinations will be grid points.

When modelling true transit transfer trips, both origins and destinations will be transit stop locations.

See also